Move FormEntryAction to own file
